+++ fabric/src/main/java/io/github/sakurawald/fuji/core/command/argument/adapter/impl/ItemStackArgumentTypeAdapter.java
@@ -5,34 +5,36 @@ import com.mojang.brigadier.context.CommandContext;
 import io.github.sakurawald.fuji.core.command.argument.adapter.abst.BaseArgumentTypeAdapter;
 import io.github.sakurawald.fuji.core.command.argument.structure.CommandArgument;
 import io.github.sakurawald.fuji.core.command.processor.CommandAnnotationProcessor;
+import lombok.SneakyThrows;
+import net.minecraft.command.argument.ItemStackArgument;
 import net.minecraft.command.argument.ItemStackArgumentType;
-import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
 import net.minecraft.server.command.ServerCommandSource;
 
 import java.util.List;
 import org.jetbrains.annotations.NotNull;
 
-public class ItemArgumentTypeAdapter extends BaseArgumentTypeAdapter {
+public class ItemStackArgumentTypeAdapter extends BaseArgumentTypeAdapter {
 
     @Override
     protected ArgumentType<?> makeArgumentType() {
         return ItemStackArgumentType.itemStack(CommandAnnotationProcessor.COMMAND_REGISTRY_ACCESS);
     }
 
+    @SneakyThrows
     @Override
     public Object makeArgumentValue(@NotNull CommandContext<ServerCommandSource> context, @NotNull CommandArgument commandArgument) {
-        return ItemStackArgumentType
-            .getItemStackArgument(context, commandArgument.getArgumentName())
-            .getItem();
+        ItemStackArgument itemStackArgument = ItemStackArgumentType.getItemStackArgument(context, commandArgument.getArgumentName());
+        return itemStackArgument.createStack(1, false);
     }
 
     @Override
     public List<Class<?>> getTypeClasses() {
-        return List.of(Item.class);
+        return List.of(ItemStack.class);
     }
 
     @Override
     public List<String> getTypeNames() {
-        return List.of("item");
+        return List.of("item", "itemstack");
     }
 }